Add files via upload
Added automatic redirects when timer runs out, variation in timer size, reworked timer output, & added field for user to indicate what url they want to go to on timeout. Method was added to make sure url has http:// & a domain, but if there is a typo it will still redirect to 404.